Option: Support omitempty
This requires changing the underlying type, which *may* be a breaking change.

AFAICT there isn't any way to do this with the struct based type, but can
be done with a slice type (or a map, but maps are heavier weight, so chose slice).

When the slice is empty (len == 0) it is considered absent, when it has an item
it is considered present.

This also allows for creating a value via:

`Option[T]{value}`

Any values appended to the slice after the first are ignored.